Remove GetPosition() method and update CheckDirection() method's parameters
Regarding the removal of GetPosition(): It is not necessary to get the position of the piece. This is because through the piece's number and the current turn (which helps identify the color we want), we can identify it in the Common. Therefore, the GetPosition() method is not required, since we already have all the values need to identify a piece, and through that, request its current position property (Pos).