Create victory verification at the end of a turn to update gameState accordingly
The new UpdateGameState() method was created in order to check for a victory at the end of every player's turn. The new CheckForWin() method was created in Controller so that the winner was not announced inside GetInput() as it was before This way, the message appears only once, with the program then waiting for player input.