Create a collection for the pieces of each color and update the IsPieceAvailable() method
These collections were created so that, using the bool indicating if the turn is for the 'Black' player, the pieces can be checked through their serial number in their respective collection. An exemple of this is that, once the player of the current turn gives the number of their piece that they want to move, it can easily be accessed and checked for its availability.